### 2017-12-19 #174 @mdsumner
Submission: tidync
Explore the contents of a NetCDF source (file or URL) presented as *variables* organized by *grids* with a database-like interface. The *hyper_filter* verb shows the effects of array-slicing expressions by value or index. Actual data read is delayed until explicitly requested, as a data frame or list of arrays.

```
Package: tidync
Title: A Tidy Approach to 'NetCDF' Data Exploration and Extraction
Version: 0.0.3
Authors@R: c(person("Michael", "Sumner", email = "mdsumner@gmail.com", role = c("aut", "cre")),
             person("Simon", "Wotherspoon", role = "ctb"),
             person("Tomas", "Remenyi", role = "ctb"),
             person("Ben", "Raymond", role = "ctb"))
Description: Tidy tools for 'NetCDF' data sources. Explore the contents of a NetCDF source (file or URL)
 presented as variables organized by grid with a database-like interface. The 'hyper_filter' verb
 shows the effects of array-slicing expressions by value or index. Actual data read is delayed until explicitly
 requested, as a data frame or list of arrays.
    forcats,
    ncdf4,
    ncmeta (>= 0.0.2),
    RNetCDF  (>= 1.9-1),
    raster,
    tabularaster,
    yesno
Suggests: ggplot2,
    memoise,
    ncdump,
    palr,
    covr
Remotes: hypertidy/ncmeta
```

https://github.com/hypertidy/tidync

- Please indicate which category or categories from our [package fit policies]

1.  **data munging** because the NetCDF model is so general that no one size fits all, tidync provides ease-of-exploration for development of domain-specific access tools as well as for understanding a new source
2.  **data extraction** because the data can be actually read by `hyper_slice` in raw form, with a higher-level output provided by the `hyper_tibble` wrapper

The target audience is users who learn so much about their particular sub-domain that they become programmers helping others in that arena. They either wrap around `tidync` to build an interface to a NetCDF source-family, or simply use it to learn to craft lower level calls more directly to the API (with packages RNetCDF, ncdf4, rgdal, rhdf5, etc).

yours differ or meet [our criteria for best-in-category]

 The dplyr `tbl_cube` is the nearest and the in-development stars does have some overlap. but I think the virtual table abstraction in tidync is novel, albeit very heavily inspired by the "laziness" of ggplot2 and the multiple-tables approach in tidygraph.  In terms of useability overall - but not in terms of generality -  the raster package is the best but it does not handle non-geographic arrays well, can only deal with 2D or 3D slices from higher forms, and does not support grids with non-regular (non-affine) coordinates.

Here's the pre-submission enquiry:

https://github.com/ropensci/onboarding/issues/167

	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

Passing locally (on Linux), but still having problems on travis.
