### 2017-12-19 da9f4d2
goodpractice cleanup
