### 2018-01-09 #174 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @mdsumner   !!

- You can put a ropensci review badge in your README

- Goodpractice output for you and reviewers to consider:

```r
It is good practice to

    R/activate.R:32:NA
    R/activate.R:33:NA
    R/activate.R:35:NA
    R/activate.R:36:NA
    R/activate.R:54:NA
    ... and 132 more lines

    R/activate.R:54:1
    R/activate.R:56:1
    R/activate.R:59:1
    R/activate.R:73:1
    R/activate.R:75:1
    ... and 53 more lines

  ✖ fix this R CMD check ERROR: Packages suggested but not
    available: ‘ncdump’ ‘palr’ The suggested packages are required for
    a complete check. Checking can be attempted without them by setting
    the environment variable _R_CHECK_FORCE_SUGGESTS_ to a false value.
    See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’
    manual.
```

Seeking reviewers now 🕐

---

Reviewers:  xx xx
Due date: xxxx
