### 2018-01-13 #174 @mdsumner
Thanks @sckott !  I added the rOpenSci badge, and I'll fix the lines and lack of tests.

I *expect* the "suggested but not available" error goes away if you install `palr` and `ncdump` from CRAN.

Editor, and reviewers: please note there's a Remotes: hypertidy/ncmeta which will need to go on CRAN soon as well.
