### 2018-01-18 #174 @sckott
Reviewers assigned

Reviewers: @timcdlucas @Nowosad
Due date: 2018-02-08
