+++ README.Rmd
@@ -53,6 +53,7 @@ tidync(file)
 There are two main ways of using tidync, interactively to explore what is there, and for extraction. The functions
 `tidync` and `activate` and `hyper_filter` allow us to hone in on the part/s of the data we want, and functions `hyper_slice` and `hyper_tibble` give raw-array and data frames with-full-coordinates forms respectively. 
 
+Also [http://www.matteodefelice.name/research/2018/01/14/tidyverse-and-netcdfs-a-first-exploration/](this blog) post by Matteo De Felice. 
 
 ### Interactive
 
