### 2018-01-28 #174 @Nowosad
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

- 4

---

- The **tidync** package provides tools to explore NetCDF files and allows to extract the data as an array, tibble, or tbl_cube.
While I think this package give ease the exploration of NetCDFs, there is still a room for improvement.
It includes, among others, expanded vignette, improved examples in the help files, more unified function names.

- I had no trouble installing/building the package from GitHub and a local clone of the repository.

#### Automated tests

- `devtools::check()` run time is rather long - ~5 minutes (Running ‘testthat.R’ [11s/211s]).
Additionally, `devtools::test()` shows that seven tests were skipped (see below).

```
══ Results ════════════════════════════════════════════════════════════════════════════════════════════════════

OK:       8
```
- The **raadtools** package is used in tests but it's not included in the Description file.
This non-CRAN package is also used in some of examples in the help files.
- `devtools::spell_check` indicates a typo in the axis_transforms.Rd file:

```
spell_check()
  WORD        FOUND IN
dimesion    axis_transforms.Rd:30
```

- The order of elements in the README file is not consistent with the [rOpenSci Packaging Guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-readme).
E.g. some badges are below the package name, while some are above the installation code.
- The README file has an appveyor badge, but the build status is unknown.
- Another packages supporting NetCDF are mentioned in the README file.
Maybe it would be worthwhile to describe how the package compares to other similar packages.
- I would suggest to rewrite the Terminology section.
There are a few problems with it, including a mix of first and third person narrative and complex stories behind the terminology used.
How about recreating it as a glossary, where each important term is explained in a few sentences?
- Also consider adding the code of conduct badge with `devtools::use_code_of_conduct`.

#### Vignette

- The **tidync** package has one vignette called "Tidy NetCDF examples".
While it give an overview of some of the package's capabilities, it also look unfinished.
- I was unable to see any difference between the first and the second example in the [tidync](https://hypertidy.github.io/tidync/articles/tidync-examples.html#tidync) section of the vignette.
It seems that `tidync(f)` and `tidync(f) %>% activate(concentration)` have the same active variable.
- The vignette also ends abruptly with a small uncommented chunk of code relating to the `hyper_tibble` function.
- I would suggest to extend the vignette (especially the data extraction section) and to add functions that are not mentioned in the vignette or README, such as `axis_transforms`, `hyper_tbl_cube`, `tidyfile`.

#### Examples

I would suggest improving some of the examples in the help files:
- The `tidyfile` and `tidync` functions examples require files from the **raadfiles** package.
- The `var_names` example looks unhelpful as it gives `NULL` as a result.

#### Function names

I found two small inconsistencies in the function names:
1. The main function is called `tidync`, next there is a large group of `hyper_*` functions, and finally there also some called differently: `activate`, `dim_names`, `var_names`.
Why this inconsistency exists?
Could the last group also have the `hyper_` prefix, e.g. `hyper_activate`?
2. `hyper_tibble()` returns `tibble`, `hyper_tbl_cube()` returns `tbl_cube`, and `hyper_slice()` returns... `array`.
Have you thought about changing the name of the last function to `hyper_array`?

#### Questions

- Interactive mode allows for subseting dimensions by values or indices.
Is it possible to get actual values (or their ranges) in the interactive mode?
It could ease the subsetting process.

@mdsumner thanks for developing the package and let me know if you have any questions regarding to my comments.
