### 2018-01-29 #174 @mdsumner
thanks @Nowosad - lots of good points here!  I'm trying a few things to come up with a better naming scheme, your feedback is very helpful

In terms of the Question "Is it possible to get actual values (or their ranges) in the interactive mode?" , I guess you mean seeing the range of "chlor_a" in this output?

```R
 library(tidync)
 f <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
 l3file <- system.file("extdata/oceandata", f, package= "tidync")
 ## filter by value
 tidync(l3file) %>% hyper_filter(lon = lon < 100)
 ```

The problem  is that this filtering is completely lazy, there's only information about the metadata (axis names and step values) - so it's effectively free in computation terms. Polling the variables for values implies quite a lot of i/o. It's possible to get the "valid range" for some variables, but I'm not sure how useful that would be (not affected by filter expressions, not always present afaik).
