### 2018-01-31 #174 @mdsumner
Ah, it didn't occur to me until now that the ranges of 1D axis variables is useful, that definitely has to be in the summary - big multi-d variables are hard to scan, but the axis ranges are readily available:

```R
x <- tidync(l3file)
axes <- axes_transforms(x)
lapply(names(axes), function(nm) range(axes[[nm]][[nm]]))
```

I couldn't see a good purrrish way to do that but I'll figure it out and make sure those ranges are in the summary print.
