### 2018-01-31 #174 @Nowosad
@mdsumner That would definitely be an improvement.
