### 2018-02-05 #174 @timcdlucas
# tidync review

Author: Tim Lucas

Estimated hours spent reviewing: 4 (lots of time wasted trying to install NetCDF)

---

#### Overview
This package provides a nice interface for dealing with NetCDF files. It seems to work as stated showing summaries of NetCDF files without reading the whole file and allowing slices from the data to be loaded and munged into a number of useful data formats.

#### Specific comments

While it could possibly be assumed that the user has NetCDF installed (why else would they want to use the package?) some details on how to install it would still be useful in the README. I have switched computers since I last used NetCDF and it did take me quite a while to get it installed.

Large block of commented out examples in `hyper_tbl_cube` and `tidyfile`.

raadtools in tidync example is a github only package.
This is quite annoying for the example of the main function.

And then I get:
```
> cf <- raadtools::cpolarfiles()
Error in UseMethod("filter_") :
  no applicable method for 'filter_' applied to an object of class "NULL"
```

ggplot2 example in `hyper_tibble` would be better as

```
> ggplot(ht %>% filter(!is.na(chlor_a)),
+        aes(x = lon, y = lat, fill = chlor_a)) + geom_tile()
```

Example in `var_names` doesn't seem to work.

```
> l3file <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
> rnc <- tidync(system.file("extdata", "oceandata", l3file, package= "tidync"))
> var_names(rnc)
NULL
```

Furthermore the description for var_names is very terse. Even though it is fairly clear what the function does just from the function name.

The tests for the function `activate` are currently commented out.

raadtools is not mentioned in `DESCRIPTION` even though it is used in many examples and unit tests.

The function `we_are_raady` is used in tests to skip. But as far as I can see the function isn't defined in `raadtools`, `raadfiles` or `tidync`.

As mentioned above, after installing `raadtools` I still get for example
```
> oisst_dayfile <- raadtools::sstfiles()$fullname[1]
Error in UseMethod("filter_") :
  no applicable method for 'filter_' applied to an object of class "NULL"
```
so I cannot run any of the unit tests.

+1 for the request for variable ranges. It is hard to know how to subset until you know this!

#### Conclusion
Overall, this looks like a promising package with a clear and useful set of functionality.
However, at the moment the documentation and tests need some work. Thanks for the work so far @mdsumner and others.
