### 2018-02-06 #174 @mdsumner
Thanks both @Nowosad and @timcdlucas I'm working through these issues, I appreciate the reviews!

One request I have is if there were NetCDF files/sources of interest that you used? Are there sources that would be compelling if I built them into the tests, or as example data in the package?  I  totally agree it's important to remove the raadtools stuff, but I otherwise find it hard to get enough interesting data to test with - I'd lost sight of how dependent on that local system I am, so apologies for that.

Would it be suitable to have the tests download files, perhaps using rappdirs and just skip those when on CRAN, but link to them in examples?
