### 2018-02-06 #174 @mdsumner
One other question, @timcdlucas were you working on Ubuntu? Is that where you have problems installing? I'm going through a fresh install on 16.04 to make sure it works, but I should do the same with Mac somehow too.
