### 2018-02-06 e44be2d
install details
