### 2018-02-06 40ffbd7
readme fixes
