+++ R/meta.R
@@ -1,106 +1,0 @@
-#' Variable names
-#' 
-#' 
-#' @param x NetCDF object
-#' @param ... currently ignored
-#'
-#' @return names of variables available
-#' @export
-#'
-#' @examples
-#' l3file <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
-#' rnc <- tidync(system.file("extdata", "oceandata", l3file, package= "tidync"))
-#' var_names(rnc)
-var_names <- function(x, ...) {
-  UseMethod("var_names")
-}
-#' @export
-#' @name var_names
-var_names.character <- function(x, ...) {
-  var_names(tidync(x))
-}
-#' @export
-#' @name var_names
-var_names.tidync <- function(x, ...) {
-  x$shape$variable
-}
-
-#' Dimension names
-#' 
-#' 
-#' @param x NetCDF object
-#' @param ... ignored
-#'
-#' @return names of available dimensions 
-#' @export
-#'
-#' @examples
-#' \dontrun{
-#' \donttest{
-#' l3file <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
-#' rnc <- tidync(system.file("extdata/oceandata", l3file, package= "tidync"))
-#' dim_names(rnc)
-#' }}
-dim_names <- function(x, ...) {
-  UseMethod("dim_names")
-}
-#' @export
-#' @name dim_names
-dim_names.character <- function(x, ...) {
-  var_names(tidync(x))
-}
-#' @export
-#' @name dim_names
-dim_names.tidync <- function(x, ...) {
-  x$variable$name
-}
-
-
-dimension_values <- function(x) {
-  .Defunct()
-  UseMethod("dimension_values")
-}
-# }
-dimension_values.tidync <- function(x) {
-  dimids <- x$grid[x$grid$grid == active(x), ] %>% inner_join(x$variable, c("variable" = "name"))
-  #dplyr::select(.data$name, .data$.variable_) %>% 
-  dimids <- dimids %>% inner_join(x$dimension,c("dimids" = "id")) %>% 
-    arrange(dimids)
-  
-  purrr::map(setNames(purrr::map(dimids$name, ~nc_get(x$file$dsn, .)), dimids$name), as_tibble)
-  
-  dimids <- dimids %>%  dplyr::inner_join(x$vardim, ".variable_") %>% 
-    dplyr::select(.data$.dimension_)
-  
-  dim_names <- x$dimension[, c("name", ".dimension_")]
-  
-  ## forcats means we maintain the right order
-  dimids %>% #dplyr::transmute(id = dimids) %>%  
-    dplyr::inner_join(x$dimension_values, ".dimension_") %>% 
-    dplyr::inner_join(dim_names, ".dimension_")
-  
-  
-}
-
-grid_dimension <- function(x) UseMethod("grid_dimension")
-grid_dimension.tidync <- function(x) {
-  x$grid %>% dplyr::filter(.data$grid == active(x)) %>% 
-    inner_join(x$variable, c("variable" = "name"))
-}
-
-variable_dimensions <- function(x) {
-  #aa <- x$variable %>% 
-  aa <- x$variable[x$variable$name == active(x), ]
-  #bb <- aa %>% 
-  #  dplyr::transmute(variable_name = .data$name, .data$.variable_) 
-  bb <- tibble::tibble(variable_name = aa$name, .variable_ = aa$.variable_)
-  cc <- bb %>% 
-    dplyr::inner_join(x$vardim, ".variable_") %>% 
-    dplyr::inner_join(x$dimension, ".dimension_")
-  dd <- tibble::tibble(variable_name = cc$variable_name, 
-                       .variable_ = cc$.variable_, 
[…]
