### 2018-02-06 55dabb0
example files, remove corruptly decompressed ifremer file
