+++ R/tidync.R
@@ -17,49 +17,67 @@ tidync <- function(x, what, ...) {
 }
 
 #' @examples
-#' #cf <- raadtools::cpolarfiles()
-#' #nc <- tidync(cf$fullname[1])
-#' #print(nc)
+#' ## a SeaWiFS (S) Level-3 Mapped (L3m) monthly (MO) chlorophyll-a (CHL)
+#' ## remote sensing product at 9km resolution (at the equator)
+#' from the NASA ocean colour group in NetCDF4 format (.nc)
+#' for 30 day period October 2009 (20092742009304) 
+#' f <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
+#' l3file <- system.file("extdata/oceandata", f, package= "tidync")
+#' tnc <- tidync(l3file)
+#' print(tnc)
+#' 
+#' ## very simple Unidata example file, with one dimension
+#' uf <- system.file("extdata/unidata", "test_hgroups.nc", package = "tidync")
+#' tidync(uf)
+#' 
+#' ## a raw grid of Southern Ocean sea ice concentration from IFREMER
+#' ## it is 12.5km resolution passive microwave concentration values
+#' ## on a polar stereographic grid, on 2 October 2017, displaying the 
+#' ## "hole in the ice" made famous here:
+#' ## https://tinyurl.com/ycbchcgn
+#' ifr <- system.file("extdata/ifremer", "20171002.nc", package = "tidync")
+#' ifrnc <- tidync(ifr)
+#' ifrnc %>% hyper_tibble(select_var = "concentration")
 #' @name tidync
 #' @export
 #' @importFrom ncmeta nc_meta
 tidync.character <- function(x, what, ...) {
   if (length(x) > 1) warning("multiple sources: only one source name allowed, ignoring all but the first")
-    fexists <- file.exists(x)
-
-   if (!fexists) cat(sprintf("not a file: \n' %s '\n\n... attempting remote connection\n", x))
-      safemeta <- purrr::safely(ncmeta::nc_meta)
-      meta <- safemeta(x)
-       if (is.null(meta$result)) {
-         cat("Oops, connection to source failed. \n")
-         print(x)
-         stop(meta$error)
-       }
-      bad_dim <- nrow(meta$result$dimension) < 1
-      bad_var <- nrow(meta$result$variable) < 1
-      if (bad_dim) {
-        warning("no dimensions found")
-      }
-      if (bad_dim) {
-        warning("no variables found")
-      }
-      if (bad_dim & bad_var) stop("no variables or dimension recognizable (is this a source with compound-types?)")
-      if (!fexists) cat("Connection succeeded. \n")      
-       meta <- meta$result
-       variable <- dplyr::mutate(meta[["variable"]], active = FALSE)
-       
-       out <- list(source = meta$source, 
-                             axis = meta$axis, 
-                             grid = meta$grid,
-                             dimension = meta$dimension, 
-                             variable = variable)
-       out$transforms <- axis_transforms(out)
-       out <- structure(out,           class = "tidync")
-       ## we can't activate nothing
-       if (nrow(out$axis) < 1) return(out)
-       if (missing(what)) what <- 1
-       out <- activate(out, what)
-
+  fexists <- file.exists(x)
+  
+  if (!fexists) cat(sprintf("not a file: \n' %s '\n\n... attempting remote connection\n", x))
+  safemeta <- purrr::safely(ncmeta::nc_meta)
+  meta <- safemeta(x)
+  if (is.null(meta$result)) {
[…]
