+++ man/dim_names.Rd
@@ -1,35 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/meta.R
-\name{dim_names}
-\alias{dim_names}
-\alias{dim_names}
-\alias{dim_names.character}
-\alias{dim_names}
-\alias{dim_names.tidync}
-\title{Dimension names}
-\usage{
-dim_names(x, ...)
-
-\method{dim_names}{character}(x, ...)
-
-\method{dim_names}{tidync}(x, ...)
-}
-\arguments{
-\item{x}{NetCDF object}
-
-\item{...}{ignored}
-}
-\value{
-names of available dimensions
-}
-\description{
-Dimension names
-}
-\examples{
-\dontrun{
-\donttest{
-l3file <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
-rnc <- tidync(system.file("extdata/oceandata", l3file, package= "tidync"))
-dim_names(rnc)
-}}
-}
