+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidync 0.0.3
 
+* dependent on ncmeta >= 0.0.2, partly to avoid crashing on invalid source/file 
+ strings
 
 * removed hyper_index and incorporated that into hyper_filter, there's now 
 only one delay-capable class which is "tidync"
