+++ man/var_names.Rd
@@ -1,32 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/meta.R
-\name{var_names}
-\alias{var_names}
-\alias{var_names}
-\alias{var_names.character}
-\alias{var_names}
-\alias{var_names.tidync}
-\title{Variable names}
-\usage{
-var_names(x, ...)
-
-\method{var_names}{character}(x, ...)
-
-\method{var_names}{tidync}(x, ...)
-}
-\arguments{
-\item{x}{NetCDF object}
-
-\item{...}{currently ignored}
-}
-\value{
-names of variables available
-}
-\description{
-Variable names
-}
-\examples{
-l3file <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
-rnc <- tidync(system.file("extdata", "oceandata", l3file, package= "tidync"))
-var_names(rnc)
-}
