+++ NAMESPACE
@@ -6,8 +6,6 @@ S3method(activate,tidync)
 S3method(active,default)
 S3method(active,tidync)
 S3method(axis_transforms,default)
-S3method(dim_names,character)
-S3method(dim_names,tidync)
 S3method(hyper_filter,tidync)
 S3method(hyper_slice,character)
 S3method(hyper_slice,tidync)
@@ -20,21 +18,17 @@ S3method(tidyfile,character)
 S3method(tidyfile,data.frame)
 S3method(tidyfile,raadfiles)
 S3method(tidync,character)
-S3method(var_names,character)
-S3method(var_names,tidync)
 export("%>%")
 export("active<-")
 export(activate)
 export(active)
 export(axis_transforms)
-export(dim_names)
 export(hyper_filter)
 export(hyper_slice)
 export(hyper_tbl_cube)
 export(hyper_tibble)
 export(tidyfile)
 export(tidync)
-export(var_names)
 importFrom(RNetCDF,close.nc)
 importFrom(RNetCDF,open.nc)
 importFrom(RNetCDF,var.get.nc)

[…]
