### 2018-02-06 #174 @timcdlucas
No problem @mdsumner.

The only time I have used NetCDF was when working on the [madingley model](https://madingley.github.io/). The files are freely downloadable from [here](https://onedrive.live.com/?authkey=%21AN5P0wXb2hRcwy4&id=3F4C6B1D2523C8D5%2123908&cid=3F4C6B1D2523C8D5). But I have no idea about permissions or anything.

I was switching between an ubuntu (16?) machine and a windows machine. I succesfully installed netcdf on ubuntu with... medium hassle. I'm not sure I actually ever succeeded on windows.

If it's useful you can test with osx on travis-ci. https://docs.travis-ci.com/user/multi-os/

And I assume you'll test with winbuilder or something before CRAN submission at least.
