### 2018-02-06 #174 @Nowosad
@mdsumner I sometimes use the E-OBS dataset - http://www.ecad.eu/download/ensembles/download.php.
