+++ NAMESPACE
@@ -14,9 +14,6 @@ S3method(hyper_tbl_cube,tidync)
 S3method(hyper_tibble,character)
 S3method(hyper_tibble,tidync)
 S3method(print,tidync)
-S3method(tidyfile,character)
-S3method(tidyfile,data.frame)
-S3method(tidyfile,raadfiles)
 S3method(tidync,character)
 export("%>%")
 export("active<-")
@@ -27,7 +24,6 @@ export(hyper_filter)
 export(hyper_slice)
 export(hyper_tbl_cube)
 export(hyper_tibble)
-export(tidyfile)
 export(tidync)
 importFrom(RNetCDF,close.nc)
 importFrom(RNetCDF,open.nc)
