### 2018-02-07 c09664f
remove raad vestiges (now in mdsumner/test.tidync)
