+++ R/utils.R
@@ -1,10 +1,3 @@
-we_are_raady <- function() {
-  fp <- getOption("default.datadir")
-  #print(fp)
-  stat <- FALSE
-  if (!is.null(fp) && file.exists(file.path(fp, "data"))) stat <- TRUE
- stat
-}
 
 # 
 # print_bytes <- function (x, digits = 3, ...) 
