### 2018-02-07 016d532
update example installed data
