+++ R/axes.R
@@ -3,7 +3,7 @@
 #' Axis 'transforms' are data frames of each dimension in a NetCDF source. 
 #' 
 #' Each transform is available by name from a list, and each data frame has
-#' the coordinate of the dimesion, its index, and a 'selected' variable
+#' the coordinate of the dimension, its index, and a 'selected' variable
 #' set by the filtering expressions in `hyper_filter` and used by the
 #' read-functions `hyper_slice` and `hyper_tibble`. 
 #' 
@@ -14,7 +14,7 @@
 #' @export
 #' @return list of axis transforms
 #' @examples 
-#' l3file <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
+#' l3file <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
 #' f <- system.file("extdata", "oceandata", l3file, package = "tidync")
 #' ax <- tidync(f) %>% axis_transforms()
 #' names(ax)
