+++ tests/testthat/test-hyper-slice.R
@@ -3,7 +3,7 @@ context("hyper-slice")
 test_that("slicing works", {
   skip_on_cran()
   l3file <- system.file("extdata/oceandata", 
-              "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc", package= "tidync")
+              "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc", package= "tidync")
   tidync(l3file) %>% hyper_filter() %>% hyper_slice()
   tidync(l3file) %>% hyper_filter() %>% hyper_tibble()
   a <- tidync(l3file) %>% hyper_filter(lon = index == 100) %>% hyper_slice()
