+++ R/hyper_slice.R
@@ -12,7 +12,7 @@
 #'
 #' @export
 #' @examples 
-#' f <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
+#' f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
 #' l3file <- system.file("extdata/oceandata", f, package= "tidync")
 #' 
 #' ## extract a raw list by filtered dimension
