+++ .travis.yml
@@ -1,50 +1,34 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
-#before_install:
-#  - sudo apt-get --yes --force-yes update -qq
-#  - sudo apt-get install -y libudunits2-dev libnetcdf-dev 
-
-
-
 language: R
 warnings_are_errors: true
-sudo: false
-cache: packages
+sudo: true
 
 addons:
     apt:
         packages:
             - libudunits2-dev
             - libnetcdf-dev
+            - netcdf-bin
 
 matrix:
     include:
         - os: linux
           r: release
           dist: trusty # 14.04
-          env: R_CODECOV=true
-          #- os: linux
-          #r: devel
-          #dist: trusty
-        - os: linux
-          r: oldrel
-          dist: trusty
+          env:
+           - BUILD_NAME=trusty_release
+           - DETAILS="release build"
         - os: osx
           r: release
-          #- os: osx
-          #r: devel
-
-# r-devel temporarily switched off; see
-# https://www.mail-archive.com/r-devel@r-project.org/msg38248.html
+          env:
+           - BUILD_NAME=osx_release
+           - DETAILS="release build"
 
 notifications:
   email:
     on_success: change
     on_failure: change
 
-r_github_packages:
-      - jimhester/covr
-
 after_success:
   - Rscript -e 'covr::codecov()'
-
