### 2018-02-08 9f2bf0d
clean up, fix tests
