+++ R/adhoc-methods.R
@@ -1,2 +1,0 @@
-plot.hyper_tbl <- function() 1
-dim.hyper_tbl <- function(x, ...)  attr(x, "dims")
