+++ tests/testthat/test-activate.R
@@ -2,7 +2,7 @@
  context("activate")
  
  test_that("various access and activation works", {
-   skip_on_cran()
+
    ufile <- system.file("extdata", "unidata", "madis-hydro.nc", package = "tidync")
    tnc <- tidync(ufile)
   ## also make sure we can active a new grid, this is the dimension ref comma separated (no brackets)
