+++ tests/testthat/test-thredds.R
@@ -1,6 +1,7 @@
 context("thredds")
 
 test_that("thredds works", {
+  skip_on_travis()
   library(tidync)
   u <- "http://coastwatch.pfeg.noaa.gov/erddap/griddap/erdQSwind3day"
   unc <- tidync(u)
