+++ DESCRIPTION
@@ -39,3 +39,4 @@ Remotes: hypertidy/ncmeta
 Roxygen: list(markdown = TRUE)
 URL: https://github.com/hypertidy/tidync
 BugReports: https://github.com/hypertidy/tidync/issues
+VignetteBuilder: knitr
