+++ appveyor.yml
@@ -15,6 +15,17 @@ cache:
 
 # Adapt as necessary starting from here
 
+
+environment:
+  global:
+    WARNINGS_ARE_ERRORS: 1
+
+  matrix:
+  - R_VERSION: release
+    R_ARCH: x64
+    BUILD_NAME: x64_release
+    DETAILS: release_build
+
 build_script:
   - travis-tool.sh install_deps
 
