+++ .travis.yml
@@ -3,6 +3,7 @@
 language: R
 warnings_are_errors: true
 sudo: true
+cache: packages
 
 addons:
     apt:
