+++ tests/testthat/test-filter.R
@@ -2,16 +2,16 @@ context("filter")
 
 
 library(dplyr)
-# test_that("standard mapped", {
-#   f1 <- system.file("extdata", "unidata", "madis-hydro.nc", package = "tidync")
-#   
-#   ## dimensions without variables
-#   tidync(f1) %>%  activate("D5,D12") %>% hyper_filter(QCcheckNum  = index < 2) %>% 
-#     expect_s3_class("tidync")
-#   tidync(f1) %>% activate("D5,D12") %>% hyper_filter() %>% 
-#     expect_s3_class("tidync")
-#   
-#   })
+test_that("unidata file", {
+  f1 <- system.file("extdata", "unidata", "madis-hydro.nc", package = "tidync")
+
+  ## dimensions without variables
+  tidync(f1) %>%  activate("D5,D12") %>% hyper_filter(QCcheckNum  = index < 2) %>%
+    expect_s3_class("tidync")
+  tidync(f1) %>% activate("D5,D12") %>% hyper_filter() %>%
+    expect_s3_class("tidync")
+
+  })
 
 test_that("indexing works", {
   l3file <- system.file("extdata/oceandata", 
