+++ NAMESPACE
@@ -20,6 +20,7 @@ export("active<-")
 export(activate)
 export(active)
 export(axis_transforms)
+export(filter.tidync)
 export(hyper_filter)
 export(hyper_slice)
 export(hyper_tbl_cube)
