+++ R/nc_get.R
@@ -2,16 +2,18 @@
 #' @importFrom ncdf4 nc_open nc_close ncvar_get
 #' @importFrom RNetCDF open.nc close.nc var.get.nc
 #' @importFrom purrr safely
-nc_get <- function(x, v) {
+nc_get <- function(x, v, test = FALSE, ...) {
   UseMethod("nc_get")
 }
-nc_get.character <- function(x, v) {
-  con <- RNetCDF::open.nc(x)
-  on.exit(RNetCDF::close.nc(con), add = TRUE)
+nc_get.character <- function(x, v, test = FALSE, ...) {
+  if (!test) {
+    con <- RNetCDF::open.nc(x)
+    on.exit(RNetCDF::close.nc(con), add = TRUE)
   
-  safe_get <- purrr::safely(nc_get.NetCDF)
-  val <- safe_get(con, v)
-  if (!is.null(val$result)) return(val$result)
+    safe_get <- purrr::safely(nc_get.NetCDF)
+    val <- safe_get(con, v)
+    if (!is.null(val$result)) return(val$result)
+  } else {
   con4 <- NULL
   con4 <- ncdf4::nc_open(x, readunlim = FALSE, verbose = FALSE, auto_GMT = FALSE, suppress_dimvals = TRUE)
   on.exit(ncdf4::nc_close(con4), add = TRUE)
@@ -23,6 +25,7 @@ nc_get.character <- function(x, v) {
     #stop("error")
     warning("dimension has no coordinates, returning index")
     return()
+  }
   }
 }
 nc_get.NetCDF <- function(x, v) {
