### 2018-02-09 7668500
each hyper_ verb now has a character method, for short circuit usage
