+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(activate,tidync)
 S3method(active,default)
 S3method(active,tidync)
 S3method(axis_transforms,default)
+S3method(hyper_filter,character)
 S3method(hyper_filter,tidync)
 S3method(hyper_slice,character)
 S3method(hyper_slice,tidync)
