### 2018-02-09 f863f82
converted expected error to expected awesome
