### 2018-02-09 b150a8d
ncmeta 0.0.2 now on CRAN
