+++ DESCRIPTION
@@ -35,7 +35,6 @@ Suggests: ggplot2,
     rmarkdown,
     testthat,
     covr
-Remotes: hypertidy/ncmeta 
 Roxygen: list(markdown = TRUE)
 URL: https://github.com/hypertidy/tidync
 BugReports: https://github.com/hypertidy/tidync/issues
