+++ docs/articles/tidync-examples.md
@@ -1,7 +1,7 @@
 ---
 title: "NetCDF examples"
 author: "Michael D. Sumner"
-date: "`r Sys.Date()`"
+date: "2018-02-10"
 output:
   rmarkdown::html_vignette:
     fig_width: 10
@@ -12,13 +12,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setup, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-options(pillar.subtle = FALSE, pillar.sigfig = 4)
-```
+
 
 The goal of tidync is to ease exploring the contents of a NetCDF source and constructing efficient 
 queries to extract arbitrary hyperslabs. The data extracted can be used directly as an array, or in 
@@ -40,10 +34,49 @@ Some conventions exist to define usage and minimal standards for metadata for pa
 Both the RNetCDF and ncdf4 packages provide a traditional summary format, familiar to many NetCDF users as the output of the command line program [`ncdump`](https://www.unidata.ucar.edu/software/netcdf/netcdf-4/newdocs/netcdf/NetCDF-Utilities.html#NetCDF-Utilities). 
 
 
-```{r}
+
+```r
 f <- system.file("extdata", "ifremer", "20171002.nc", package = "tidync")
 library(RNetCDF)
 print.nc(open.nc(f))
+#> dimensions:
+#>         ni = 632 ;
+#>         nj = 664 ;
+#>         time = 1 ;
+#> variables:
+#>         int time(time) ;
+#>                 time:long_name = "time" ;
+#>                 time:units = "hours since 1900-1-1 0:0:0" ;
+#>         byte concentration(ni, nj, time) ;
+#>                 concentration:long_name = "sea-ice concentration" ;
+#>                 concentration:units = "percent" ;
+#>                 concentration:scale_factor = 1 ;
+#>                 concentration:add_offset = 0 ;
+#>                 concentration:missing_value = -128 ;
+#>                 concentration:_FillValue = -128 ;
+#>         byte quality_flag(ni, nj, time) ;
+#>                 quality_flag:long_name = "quality_flag" ;
+#>                 quality_flag:units = "n/a" ;
+#>                 quality_flag:scale_factor = 1 ;
+#>                 quality_flag:add_offset = 0 ;
+#>                 quality_flag:missing_value = -128 ;
+#>                 quality_flag:_FillValue = -128 ;
+#> 
+#> // global attributes:
+#>                 :CONVENTIONS = "COARDS" ;
+#>                 :long_name = "Sea-ice concentration as observed by SSM/I" ;
+#>                 :short_name = "PSI-F18-Concentration" ;
+#>                 :producer_agency = "IFREMER" ;
+#>                 :producer_institution = "CERSAT" ;
+#>                 :netcdf_version_id = "3.4" ;
+#>                 :product_version = "2.0" ;
+#>                 :creation_time = "2017-280T08:26:34.000" ;
+#>                 :time_resolution = "daily" ;
+#>                 :grid = "NSIDC" ;
+#>                 :pole = "south" ;
+#>                 :spatial_resolution = "12.5 km" ;
+#>                 :platform_id = "F18" ;
+#>                 :instrument = "SSM/I" ;
 ```
 
 Using `ncdump` at the command line on a suitable system would yield very similar output to the print above.  
@@ -73,10 +106,26 @@ On first contact with the file, the available variables are classified by grid a
 dimension.  The "active" grid is the one that queries may be made against, and may be changed with the `activate` function. 
 
 
-```{r}
+
+```r
 library(tidync)
 tidync(f)
-
+#> 
[…]
