### 2018-03-06 #174 @mdsumner
I have been delayed, but yes there is progress!  I'm doing it in this branch:
 https://github.com/hypertidy/tidync/commits/ropensci-review

(most other branches will likely be deleted but I need to check a few things)

* major clean up of examples, so no raad-only local stuff
* clarified some of the verb problems, will describe
* print now includes source vs. filtered length/start/count info, needs to be separated for active vs. available dimensions
* identified some inconsistencies that I need to sort out
* experimented with some new ways of dealing with lazy-data, that are promising but shouldn't delay this review any further

I created a target release for post review v0.0.4 https://github.com/hypertidy/tidync/issues?q=is%3Aissue+is%3Aopen+label%3Av0.0.4

There's nothing specific I can point to for help at this point, so I appreciate the patience. Hoping to push forward again next week.
