### 2018-03-28 d9d2a74
Update hyper_slice.R
