### 2018-04-16 565b06c
slice doc
