+++ R/tidync-package.r
@@ -9,7 +9,7 @@
 #' \tabular{ll}{
 #'  \code{\link{tidync}} \tab extracts all file metadata in an object useable by other tidync functions\cr
 #'  \code{\link{hyper_filter}} \tab processes user-driven expressions to filter based on NetCDF dimensions \cr
-#'  \code{\link{hyper_slice}} \tab extracts a raw data array based on a NetCDF index \cr
+#'  \code{\link{hyper_array}} \tab extracts a raw data array based on a NetCDF index \cr
 #'  \code{\link{hyper_tibble}} \tab extracts the slice as a data frame with all dimension values \cr
 #'  }
 #'  The scheme generally processes dimension filters into NetCDF extraction indexes and these are always available
