### 2018-08-28 360db90
group by only for polygons
