### 2018-08-29 e4568ff
group by sf with shape = arg
