### 2018-09-13 78a2806
make transforms accessible from raw array output
