### 2018-09-13 d124332
internal raster constructor
