### 2019-01-10 1af4e96
handle dim vars that are character
