+++ NAMESPACE
@@ -40,7 +40,9 @@ importFrom(dplyr,group_by)
 importFrom(dplyr,inner_join)
 importFrom(dplyr,mutate)
 importFrom(dplyr,row_number)
+importFrom(dplyr,summarise)
 importFrom(dplyr,summarize)
+importFrom(dplyr,ungroup)
 importFrom(forcats,as_factor)
 importFrom(magrittr,"%>%")
 importFrom(ncdf4,nc_close)
