### 2019-02-13 3217782
protect against choosing NC_CHAR vars by default
