### 2019-02-14 55d48f2
better default var choice
