+++ man/dplyr-verbs.Rd
@@ -30,7 +30,9 @@ tbl_vars.tidync(x)
 Very experimental
 }
 \examples{
-x <- tidync(raadtools::sshfiles()$fullname[1])
- x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% group_by(longitude) \%>\%   summarize(adt = mean(adt, na.rm = TRUE))
- x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% group_by(longitude) \%>\%   summarize_all(mean, na.rm = TRUE)
+\dontrun{
+#x <- tidync(raadtools::sshfiles()$fullname[1])
+#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% group_by(longitude) \%>\%   summarize(adt = mean(adt, na.rm = TRUE))
+#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% group_by(longitude) \%>\%   summarize_all(mean, na.rm = TRUE)
+}
 }
