+++ man/dplyr-verbs.Rd
@@ -22,6 +22,8 @@ tbl_vars.tidync(x)
 
 \item{...}{passed into dplyr engine}
 
+\item{.x}{tidync object}
+
 \item{add}{add to existing groups}
 
 \item{shape}{experimental tidync}
@@ -32,7 +34,9 @@ Very experimental
 \examples{
 \dontrun{
 #x <- tidync(raadtools::sshfiles()$fullname[1])
-#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% group_by(longitude) \%>\%   summarize(adt = mean(adt, na.rm = TRUE))
-#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% group_by(longitude) \%>\%   summarize_all(mean, na.rm = TRUE)
+#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% 
+# group_by(longitude) \%>\%   summarize(adt = mean(adt, na.rm = TRUE))
+#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% 
+#group_by(longitude) \%>\%   summarize_all(mean, na.rm = TRUE)
 }
 }
