+++ .travis.yml
@@ -14,6 +14,8 @@ addons:
             - libudunits2-dev
             - libnetcdf-dev
             - netcdf-bin
+            - libproj-dev 
+
              
 matrix:
     include:
