### 2019-02-25 #174 @sckott
👋  @mdsumner just a heads up that our policies https://ropensci.github.io/dev_guide/policies.html#review-process state that issues will be closed after one year from applying the holding label (mar 2019, next month) and you'll need to resubmit if the issue is closed
