### 2019-02-25 1d1743b
better doc for https://github.com/hypertidy/tidync/issues/63
