+++ README.Rmd
@@ -23,14 +23,12 @@ knitr::opts_chunk$set(
 [![Coverage status](https://codecov.io/gh/hypertidy/tidync/branch/master/graph/badge.svg)](https://codecov.io/github/hypertidy/tidync?branch=master)
 
 
-
-
 The goal of tidync is to ease exploring the contents of a NetCDF source and constructing efficient 
 queries to extract arbitrary hyperslabs. The data extracted can be used directly as an array, or in 
 "long form" form as a data frame for "tidy" analysis and visualization contexts. 
 
 NetCDF is **Network Common Data Form** a very common, and very general way to store and work with 
-scientific array-based data. NetCDF is defined and provided by [Unidata](https://www.unidata.ucar.edu/software/netcdf/). R has (independent) support for NetCDF via the [RNetCDF](https://CRAN.R-project.org/package=RNetCDF), [ncdf4](https://CRAN.R-project.org/package=ncdf4), [rhdf5](https://bioconductor.org/packages/release/bioc/html/rhdf5.html) and [rgdal](https://CRAN.R-project.org/package=rgdal) packages. 
+scientific array-based data. NetCDF is defined and provided by [Unidata](https://www.unidata.ucar.edu/software/netcdf/). R has (independent) support for NetCDF via the [RNetCDF](https://CRAN.R-project.org/package=RNetCDF), [ncdf4](https://CRAN.R-project.org/package=ncdf4), [rhdf5](https://bioconductor.org/packages/release/bioc/html/rhdf5.html), [stars](https://CRAN.R-project.org/package=stars),  [sf](https://CRAN.R-project.org/package=sf) and [rgdal](https://CRAN.R-project.org/package=rgdal) packages. 
 
 This project uses RNetCDF for the primary access to the NetCDF library. 
 
@@ -40,8 +38,8 @@ This project uses RNetCDF for the primary access to the NetCDF library.
 You can install tidync from github with:
 
 ```{r gh-installation, eval = FALSE}
-# install.packages("devtools")
-devtools::install_github("hypertidy/tidync", dependencies = TRUE)
+# install.packages("remotes")
+remotes::install_github("hypertidy/tidync", dependencies = TRUE)
 ```
 
 Also required are packages 'ncdf4' and 'RNetCDF', so first make sure you can install and use these. 
@@ -67,7 +65,6 @@ On Linux you will need at least the following installed by an administrator,
 here tested on Ubuntu Xenial 16.04. As far as I know nothing extra above 
 the libnetcdf-dev library is required for OpenDAP/Thredds. 
 
-I otherwise recommend using docker, with for example [rocker/geospatial](https://github.com/rocker-org/geospatial).
 
 ```bash
 apt update 
