+++ man/hyper_tibble.Rd
@@ -45,6 +45,5 @@ lapply(hyper_array(f, lat = lat > 0, lon = index > 3000), dim)
 ht   
 library(ggplot2)
 ggplot(ht \%>\% filter(!is.na(chlor_a)), 
-aes(x = lon, y = lat, fill = chlor_a)) + geom_point()
-
+aes(x = lon, y = lat, fill = chlor_a)) + geom_tile()
 }
