### 2019-02-27 cc40fdc
fix woeful dimension mismatch
