### 2019-02-27 9142f62
NC_CHAR!
