+++ NEWS.md
@@ -1,5 +1,7 @@
 # dev
 
+* Now supports 'NC_CHAR' type, by exploding these into the array size expected. 
+
 * Breaking change: when using `tidync$grid`it's now expected that this must 
   be `tidyr::unnest()`ed in order to expand out the grid list per variable, in line with
   https://github.com/hypertidy/ncmeta/issues/26. 
