+++ tests/testthat/test-files-sources.R
@@ -11,8 +11,20 @@ test_that("files and bad files are handled", {
   expect_warning(try(tidync(l3b_file), silent = TRUE), "no dimensions found")
   expect_warning(try(tidync(l3b_file), silent = TRUE), "no variables found")
   expect_warning(try(tidync(l3b_file), silent = TRUE), "no variables recognizable")  
+  f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"  
+  l3file <- system.file("extdata/oceandata", f, package= "tidync", mustWork = TRUE)
   
-
+  expect_warning(try(tidync(l3file[c(1, 1)])), "multiple sources: only one source name allowed, ignoring all but the first")
+  tfile <- tempfile()
+  nothingfile <- RNetCDF::create.nc(tfile)
+  RNetCDF::close.nc(nothingfile)
+  
+  expect_warning(ouch <- try(tidync::tidync(tfile), silent = TRUE), sample(c("no variables recognizable", "no dimensions found", "no variables found"), 1))
+  
+  
+  tnc <- tidync(l3file)
+  tnc$grid <- tnc$grid[0, ]
+  expect_output(print(tnc), "No recognizable dimensions or variables")
 })
 
 test_that("verbs have a character method", {
