### 2019-02-27 2f78e26
separate active dimensions in print https://github.com/hypertidy/tidync/issues/60
