+++ tests/testthat/test-files.R
@@ -1,7 +1,6 @@
-# context("files")
-# 
-# fname <- paste(sample(unlist(strsplit("somecrazyfile", ""))), collapse = "")
-# test_that("file not found is friendly", {
-#   #expect_error(tidync(fname), "No such file or directory")
-#   expect_message(tidync(fname), "cannot read from")
-# })
+context("files")
+
+fname <- paste(sample(unlist(strsplit("somecrazyfile", ""))), collapse = "")
+test_that("file not found is friendly", {
+  expect_error(tidync(fname), "failed to open")
+})
