+++ man/hyper_filter.Rd
@@ -20,7 +20,7 @@ data frame
 \description{
 The 'hyper filter' acts on a 'tidync' object by matching one or more filtering
 expressions like with \code{dplyr::filter}. This allows us to lazily specify a subset
-from a NetCDF array without pulling  any datadata. The modified object may be printed
+from a NetCDF array without pulling  any data. The modified object may be printed
 to see the effects of subsetting, or saved for further use.
 }
 \details{

[…]
