+++ tests/testthat/test-nc_char.R
@@ -4,6 +4,6 @@ f <- system.file("extdata/argo/MR5905167_372.nc", package = "tidync", mustWork =
 test_that("nc_char is handled gracefully", {
   tnc <- tidync(f)
   expect_s3_class(tnc$variable, "tbl_df")
-  expect_true(dim(tnc$variable) == c(77L, 7L))
+  expect_true(all(dim(tnc$variable) == c(77L, 7L)))
   expect_output(print(tnc))
 })
