+++ README.md
@@ -3,21 +3,27 @@
 
 # tidync
 
-[![](https://badges.ropensci.org/174_status.svg)](https://github.com/ropensci/onboarding/issues/174)
-[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/tidync)](https://cran.r-project.org/package=tidync)
 [![Travis-CI Build
-Status](http://badges.herokuapp.com/travis/hypertidy/tidync?branch=master&env=BUILD_NAME=trusty_release&label=linux)](https://travis-ci.org/hypertidy/tidync)
-[![Build
+Status](http://badges.herokuapp.com/travis/hypertidy/tidync?branch=master&env=BUILD_NAME=trusty_release&label=linux)](https://travis-ci.org/hypertidy/tidync)[![Build
 Status](http://badges.herokuapp.com/travis/hypertidy/tidync?branch=master&env=BUILD_NAME=osx_release&label=osx)](https://travis-ci.org/hypertidy/tidync)
 [![AppVeyor Build
 Status](https://ci.appveyor.com/api/projects/status/github/hypertidy/tidync?branch=master&svg=true)](https://ci.appveyor.com/project/mdsumner/tidync-3m5h7)
 [![Coverage
-status](https://codecov.io/gh/hypertidy/tidync/branch/master/graph/badge.svg)](https://codecov.io/github/hypertidy/tidync?branch=master)
+status](https://codecov.io/gh/hypertidy/tidync/branch/master/graph/badge.svg)](https://codecov.io/github/hypertidy/tidync?branch=master)[![](https://badges.ropensci.org/174_status.svg)](https://github.com/ropensci/onboarding/issues/174)[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/tidync)](https://cran.r-project.org/package=tidync)
 
 The goal of tidync is to ease exploring the contents of a NetCDF source
 and constructing efficient queries to extract arbitrary hyperslabs. The
 data extracted can be used directly as an array, or in long-form data
-frame.
+frame. In contrast to other packages tidync helps reduce the volume of
+code required to discover and read the contents of NetCDF, with simple
+steps:
+
+  - Connect and summarize `tidync()`.
+  - (optionally) Specify source variables `activate()`.
+  - (optionally) Specify array sub-setting (slicing) `hyper_filter()`.
+  - Read array data in either native `hyper_array()` or long-form
+    (`hyper_tibble()`).
 
 NetCDF is **Network Common Data Form** a very common, and very general
 way to store and work with scientific array-based data. NetCDF is
@@ -237,76 +243,12 @@ explored that here and it may or may not be a good idea:
 
 ## Terminology
 
-Here I use the term “slab” as a generalized “array” (in the R sense)
-that may be read from a NetCDF. We must provide the NetCDF API with a
-“slab index”, i.e. both a start and a count vector each the same
-length (the same as the number of dimensions as the array variable),
-that is the only way to read them.
-
-In R terms a 3D array would be indexed like
-
-``` r
-arr[1:10, 2:12, 3:5]
-```
-
-and that would be analogous to
-
-``` r
-ncvar_get(con, start = c(1, 2, 3), count = c(10, 11, 3))
-```
-
-If we only wanted a “sparse trace” through the array in R we can do
-
-``` r
-arr[cbind(c(2, 4), c(5, 6), c(3, 4)]
-```
-
-which would pull out 2-values from 2 arbitrary positions. The API
[…]
