### 2019-02-28 6dba0bb
remove dimension id from print (implicit in 'dim') and active field
