### 2019-02-28 a90ed76
tidy up for 0.4.0 candidate (response to rOpenSci  reviews)
