+++ vignettes/tidync-examples.md
@@ -1,0 +1,462 @@
+---
+title: "NetCDF examples"
+author: "Michael D. Sumner"
+date: "2019-02-28"
+output:
+  rmarkdown::html_vignette:
+    fig_width: 10
+    fig_height: 10
+vignette: >
+  %\VignetteIndexEntry{Tidy NetCDF examples}
+  %\VignetteEngine{knitr::knitr}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+
+The goal of tidync is to ease exploring the contents of a NetCDF source and constructing efficient 
+queries to extract arbitrary hyperslabs. The data extracted can be used directly as an array, or in 
+"long form" form as a data frame for "tidy" analysis and visualization contexts. 
+
+NetCDF is **Network Common Data Form**, a [data model][not-a-format] and an [API](https://en.wikipedia.org/wiki/Application_programming_interface).
+
+This is a very common, and *very general* way to store and work with 
+scientific array-based data. NetCDF is defined and provided by [Unidata](https://www.unidata.ucar.edu/software/netcdf/).
+
+Here we introduce traditional concepts of NetCDF, and show examples with built-in files and online sources to demonstrate tidync functionality. 
+
+# NetCDF
+
+NetCDF is a very widely used file format for storing array-based data as *variables*. The **space** occupied by a **variable** is defined by its **dimensions** and their metadata. Dimensions are by definition *one-dimensional* (i.e. an atomic vector, in R) consisting of one or more elements, a rectilinear virtual array with coordinate metadata on its units, type and interpretation. The **space** of a variable is defined as one or more of the dimensions in the file, but a variable won't necessarily use all the available dimensions and no dimensions are mandatory or particularly special. 
+
+Some conventions exist to define usage and minimal standards for metadata for particular file schemas, but these are many and varied, and not always adhered to. 
+
+The R community is not particularly strong with use of NetCDF, though it
+is common and widely used it pales compared to use in general climate science work, and there the most used tool is the [CDO Climate Data Operators](https://code.mpimet.mpg.de/projects/cdo).  In R the most common
+tools used are ncdf4 and raster (which uses ncdf4). 
+
+Both the RNetCDF and ncdf4 packages provide a traditional summary format, familiar to many NetCDF users as the output of the command line program [`ncdump`](https://www.unidata.ucar.edu/software/netcdf/netcdf-4/newdocs/netcdf/NetCDF-Utilities.html#NetCDF-Utilities). 
+
+
+
+```r
+f <- system.file("extdata", "ifremer", "20171002.nc", package = "tidync")
+library(RNetCDF)
+print.nc(open.nc(f))
+#> dimensions:
+#>         ni = 632 ;
+#>         nj = 664 ;
+#>         time = 1 ;
+#> variables:
+#>         int time(time) ;
+#>                 time:long_name = "time" ;
+#>                 time:units = "hours since 1900-1-1 0:0:0" ;
+#>         byte concentration(ni, nj, time) ;
+#>                 concentration:long_name = "sea-ice concentration" ;
+#>                 concentration:units = "percent" ;
+#>                 concentration:scale_factor = 1 ;
+#>                 concentration:add_offset = 0 ;
+#>                 concentration:missing_value = -128 ;
+#>                 concentration:_FillValue = -128 ;
[…]

[…]
