### 2019-02-28 #174 @mdsumner
Hi!  Thanks so much for the reviews, the package has been greatly improved by them!  Very sorry for the huge delay.

My response is pretty terse but I believe I've covered the vast majority of issues.  I've added a todo list at the bottom for things I know I haven't finished addressing.

@Nowosad

* check time is now ~1.5 min, and test coverage improved
* no raadtools used
* spelling check improved (WORDLIST added)
* badges improved
* clarifications within readme
* short terminology list added
* clarified the vignette
* renamed hyper_slice to hyper_array
* subset values are now seen in the tidync print out (min/max and corresponding dmin/dmax)

Some tests have warnings, but I need to explore that more deeply and expect the CRAN
process will require it to be finalized.

-----
@timcdlucas

* installation instructions greatly improved
* automated testing improved
* cleaned up examples, no raadtools
* used geom_tile() in ggplot2 example
* var_names() removed, now use hyper_vars() to return information about variables
and hyper_dims() about dimensions (note example file changed)
* improved linkage of help pages of tidnyc, print.tidync, hyper_dims and hyper_vars.

Example code, relevant to `hyper_vars` query  (was `var_names`):

``` r
library(tidync)
l3file <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
rnc <- tidync(system.file("extdata", "oceandata", l3file, package= "tidync"))
hyper_vars(rnc)
#> # A tibble: 1 x 6
#>      id name    type     ndims natts dim_coord
#>   <dbl> <chr>   <chr>    <dbl> <dbl> <lgl>
#> 1     0 chlor_a NC_FLOAT     2    10 FALSE
hyper_dims(rnc)
#> # A tibble: 2 x 7
#>   name  length start count    id unlim coord_dim
#>   <chr>  <dbl> <int> <int> <dbl> <lgl> <lgl>
#> 1 lon     4320     1  4320     1 FALSE TRUE
#> 2 lat     2160     1  2160     0 FALSE TRUE
```

The remaining todos:

- [ ] finalize vignette (along with any inconsistent funs)
- [ ] update ncmeta and tabularaster on CRAN
- [ ] align to goodpractice
