### 2019-02-28 e3a9469
remove unused dplyr extensions
