+++ man/dplyr-verbs.Rd
@@ -1,42 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/filter-groupby.R
-\name{tbl_vars.tidync}
-\alias{tbl_vars.tidync}
-\alias{summarise.tidync}
-\alias{group_by.tidync}
-\alias{ungroup.tidync}
-\title{Dplyr verbs for tidync}
-\usage{
-tbl_vars.tidync(x)
-
-\method{summarise}{tidync}(.data, ...)
-
-\method{group_by}{tidync}(.x, ..., add = FALSE, shape = NULL)
-
-\method{ungroup}{tidync}(x, ...)
-}
-\arguments{
-\item{x}{tidync object}
-
-\item{.data}{tidync object}
-
-\item{...}{passed into dplyr engine}
-
-\item{.x}{tidync object}
-
-\item{add}{add to existing groups}
-
-\item{shape}{experimental tidync}
-}
-\description{
-Very experimental
-}
-\examples{
-\dontrun{
-#x <- tidync(raadtools::sshfiles()$fullname[1])
-#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% 
-# group_by(longitude) \%>\%   summarize(adt = mean(adt, na.rm = TRUE))
-#  x \%>\% hyper_filter(longitude = longitude > 147, latitude = latitude < - 30) \%>\% 
-#group_by(longitude) \%>\%   summarize_all(mean, na.rm = TRUE)
-}
-}
