+++ README.Rmd
@@ -17,11 +17,12 @@ knitr::opts_chunk$set(
 
 [![Travis-CI Build Status](http://badges.herokuapp.com/travis/hypertidy/tidync?branch=master&env=BUILD_NAME=trusty_release&label=linux)](https://travis-ci.org/hypertidy/tidync)[![Build Status](http://badges.herokuapp.com/travis/hypertidy/tidync?branch=master&env=BUILD_NAME=osx_release&label=osx)](https://travis-ci.org/hypertidy/tidync) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/hypertidy/tidync?branch=master&svg=true)](https://ci.appveyor.com/project/mdsumner/tidync-3m5h7) [![Coverage status](https://codecov.io/gh/hypertidy/tidync/branch/master/graph/badge.svg)](https://codecov.io/github/hypertidy/tidync?branch=master)[![](https://badges.ropensci.org/174_status.svg)](https://github.com/ropensci/onboarding/issues/174)[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing) [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/tidync)](https://cran.r-project.org/package=tidync)
 
-
-
 The goal of tidync is to ease exploring the contents of a NetCDF source and constructing efficient 
-queries to extract arbitrary hyperslabs. The data extracted can be used directly as an array, or in 
-long-form data frame. In contrast to other packages tidync helps reduce the volume of code required to discover and read the contents of NetCDF, with simple steps: 
+queries to extract arbitrary sub-arrays. 
+
+The data extracted can be used directly as an array, or in long-form data frame.
+In contrast to other packages tidync helps reduce the volume of code required to
+discover and read the contents of NetCDF, with simple steps:
 
 * Connect and summarize `tidync()`.
 * (optionally) Specify source variables `activate()`.
@@ -29,15 +30,11 @@ long-form data frame. In contrast to other packages tidync helps reduce the volu
 * Read array data in either native `hyper_array()` or long-form `hyper_tibble()`. 
 
 NetCDF is **Network Common Data Form** a very common, and very general way to store and work with 
-scientific array-based data. NetCDF is defined and provided by [Unidata](https://www.unidata.ucar.edu/software/netcdf/). R has (independent) support for NetCDF via the [RNetCDF](https://CRAN.R-project.org/package=RNetCDF), [ncdf4](https://CRAN.R-project.org/package=ncdf4), [rhdf5](https://bioconductor.org/packages/release/bioc/html/rhdf5.html), [stars](https://CRAN.R-project.org/package=stars),  [sf](https://CRAN.R-project.org/package=sf) and [rgdal](https://CRAN.R-project.org/package=rgdal) packages. 
+scientific array-based data. NetCDF is defined and provided by [Unidata](https://www.unidata.ucar.edu/software/netcdf/). R has (independent) support for NetCDF via the [ncdf4](https://CRAN.R-project.org/package=ncdf4), [rhdf5](https://bioconductor.org/packages/release/bioc/html/rhdf5.html), [RNetCDF](https://CRAN.R-project.org/package=RNetCDF), [rgdal](https://CRAN.R-project.org/package=rgdal), [sf](https://CRAN.R-project.org/package=sf) and [vapour](https://CRAN.R-project.org/package=vapour) packages. 
 
-
-This project uses RNetCDF for the primary access to the NetCDF library. 
-
-
[…]
