### 2019-02-28 #174 @sckott
thanks very much @mdsumner for the responses. i've removed the holding label. For the remaining to do's any guess on the timeline for those?

@Nowosad @timcdlucas if you have any further questions/comments on the response do chime in here
