+++ tests/testthat/test-activate.R
@@ -40,3 +40,5 @@
   expect_warning(activate(1, "a"))
 
 })
+
+ 
