+++ tests/testthat/test-hyper-slice.R
@@ -18,12 +18,16 @@ test_that("slicing works", {
   expect_error(
     ## 
     tidync(l3file) %>% hyper_filter(lat = between(lat, -42, -45), 
-                                               lon = lon > 170) %>% hyper_array()
+                                               lon = lon > 170) %>% 
+      hyper_array()
   )
   
-  ufile <- system.file("extdata", "unidata", "madis-hydro.nc", package = "tidync")
+  ufile <- system.file("extdata", "unidata", "madis-hydro.nc", 
+                       package = "tidync")
   a <- hyper_array(tidync(ufile) %>% activate("D5,D12"))
-  expect_named(a, c("precip5minQCD", "precip1hrQCD", "precip3hrQCD", "precip6hrQCD", "precip12hrQCD", "precip24hrQCD", "precipAccumQCD"))
+  expect_named(a, c("precip5minQCD", "precip1hrQCD", "precip3hrQCD", 
+                    "precip6hrQCD", "precip12hrQCD", "precip24hrQCD", 
+                    "precipAccumQCD"))
   a <- hyper_array(tidync(ufile) %>% activate(precip3hrQCD))
   expect_named(a, c("precip3hrQCD"))
 })
