### 2019-03-03 #174 @mdsumner
I've pushed version 0.1.0 to the repo with  [this commit](https://github.com/hypertidy/tidync/commit/5fb11a43913f11bd2a20be04c3e4836986b2fdb9)

I consider it review-reply ready now. (ncmeta alignment for CRAN will be tricky as I have to sort out version-robustness for stars).  It's not passing on travis, but it is passing locally so I'm hoping it's a temporary issue there.

I'm struggling to finish that vignette, so I'm just letting you know it still ends awkwardly.  I have to leave this for a week or two, so I can't get to that.

The testing and documentation now is much better I believe. The set of functions is clearer, and are now summarized in `?tidync-package`:

https://hypertidy.github.io/tidync/reference/tidync-package.html

(That page btw is built from source files longer than 80char, but I can't see how to shorten them and still work?)

Thanks again!
