+++ .travis.yml
@@ -24,7 +24,7 @@ matrix:
           dist: trusty # 14.04
           env:
            - BUILD_NAME=trusty_release
-           - DETAILS="old build"
+           - DETAILS="old trusty"
         - os: linux
           r: release
           dist: xenial # 16.04
