### 2019-03-04 #174 @sckott
thanks @timcdlucas

@Nowosad when you get a chance can you look at the new version and see if you have any more feedback?
