+++ R/hyper_tbl_cube.R
@@ -16,8 +16,17 @@
 #' activate(chlor_a), lon = lon > 107, lat = abs(lat) < 30))
 #' ufile <- system.file("extdata", "unidata", "test_hgroups.nc", 
 #'  package = "tidync", mustWork = TRUE)
-#' tidync(ufile)
-#' hyper_tbl_cube(tidync(ufile))
+#'  
+#' ## some versions of NetCDF don't support this file
+#' ## (4.1.3 https://github.com/hypertidy/tidync/issues/82)
+#' group_nc <- try(tidync(ufile), silent = TRUE)
+#' if (!inherits(group_nc, "try-error")) {
+#'  res <-  hyper_tbl_cube(tidync(ufile))
+#'  print(res)
+#' } else {
+#'  ## the error was
+#'  writeLines(c(group_nc))
+#' }
 #' @return `dplyr::tbl_cube`
 #' @export
 hyper_tbl_cube <- function(x, ...) {
