+++ man/hyper_tbl_cube.Rd
@@ -35,6 +35,15 @@ l3file <- system.file("extdata/oceandata", f, package= "tidync")
 activate(chlor_a), lon = lon > 107, lat = abs(lat) < 30))
 ufile <- system.file("extdata", "unidata", "test_hgroups.nc", 
  package = "tidync", mustWork = TRUE)
-tidync(ufile)
-hyper_tbl_cube(tidync(ufile))
+ 
+## some versions of NetCDF don't support this file
+## (4.1.3 https://github.com/hypertidy/tidync/issues/82)
+group_nc <- try(tidync(ufile), silent = TRUE)
+if (!inherits(group_nc, "try-error")) {
+ res <-  hyper_tbl_cube(tidync(ufile))
+ print(res)
+} else {
+ ## the error was
+ writeLines(c(group_nc))
+}
 }
