### 2019-03-05 #174 @Nowosad
I've already took a look at the changes and they are :+1: . The only thing that can be improved is the vignette.
