### 2019-03-14 #174 @timcdlucas
Hi all,

Great work @mdsumner.

Hope I'm looking at the right version. I'm using bfb975ce565fc807e482497c7a7b953b6e66463c

I still get 1 test failure.

```
>   expect_equal(active(tnc1),  "D1,D12")

> active(tnc2)
[1] "D1,D12"
```

in `test-activate.R`.

Other than that the only thing I can see is that the readme isn't currently runeable. `filename` is not defined for example. This chunk is set as `eval = FALSE` so I guess it's deliberate, but I think it's more useful if the readme is something that can just be copied into an R session.

Otherwise I'm happy!

Tim
