### 2019-03-18 ca6f7fb
update vignette and docs in response to review
