+++ README.md
@@ -134,22 +134,22 @@ tidync(file)
 #>   
 #>   dim   name  length    min   max start count   dmin  dmax unlim coord_dim 
 #>   <chr> <chr>  <dbl>  <dbl> <dbl> <int> <int>  <dbl> <dbl> <lgl> <lgl>     
-#> 1 D0    lat     2160 -180.  180.      1  2160 -180.  180.  FALSE TRUE      
-#> 2 D1    lon     4320  -90.0  90.0     1  4320  -90.0  90.0 FALSE TRUE      
+#> 1 D0    lat     2160  -90.0  90.0     1  2160  -90.0  90.0 FALSE TRUE      
+#> 2 D1    lon     4320 -180.  180.      1  4320 -180.  180.  FALSE TRUE      
 #>   
 #> Inactive dimensions:
 #>   
 #>   dim   name          length   min   max unlim coord_dim 
 #>   <chr> <chr>          <dbl> <dbl> <dbl> <lgl> <lgl>     
-#> 1 D2    rgb                3     1   256 FALSE FALSE     
-#> 2 D3    eightbitcolor    256     1     3 FALSE FALSE
+#> 1 D2    rgb                3     1     3 FALSE FALSE     
+#> 2 D3    eightbitcolor    256     1   256 FALSE FALSE
 ```
 
 There are two main ways of using tidync, interactively to explore what
 is there, and for extraction. The functions `tidync` and `activate` and
 `hyper_filter` allow us to hone in on the part/s of the data we want,
-and functions `hyper_array` and `hyper_tibble` give raw-array and data
-frames with-full-coordinates forms respectively.
+and functions `hyper_array`, `hyper_tibble` and `hyper_tbl_cube` give
+raw-array or data frames.
 
 Also
 [http://www.matteodefelice.name/research/2018/01/14/tidyverse-and-netcdfs-a-first-exploration/](this%20blog)
@@ -165,17 +165,190 @@ filter it by its
 index).
 
 ``` r
+filename <- system.file("extdata/argo/MD5903593_001.nc", package = "tidync")
 ## discover the available entities, and the active grid's dimensions and variables
 tidync(filename)
+#> 
+#> Data Source (1): MD5903593_001.nc ...
+#> 
+#> Grids (16) <dimension family> : <associated variables> 
+#> 
+#> [1]   D0,D9,D11,D8 : SCIENTIFIC_CALIB_DATE
+#> [2]   D6,D9,D11,D8 : PARAMETER
+#> [3]   D7,D9,D11,D8 : SCIENTIFIC_CALIB_EQUATION, SCIENTIFIC_CALIB_COEFFICIENT, SCIENTIFIC_CALIB_COMMENT
+#> [4]   D6,D9,D8     : STATION_PARAMETERS
[…]
