### 2019-03-18 dd6f3bd
tighten documentation
