+++ R/hyper_tibble.R
@@ -19,6 +19,8 @@
 #' @export
 #' @importFrom dplyr %>%
 #' @export %>%
+#' @seealso [hyper_array()] and [hyper_tbl_cube()]  which are also delay-breaking 
+#' functions that cause data to be read 
 #' @examples
 #' l3file <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
 #' f <- system.file("extdata", "oceandata", l3file, package= "tidync")
