+++ R/hyper_vars_dims.R
@@ -6,7 +6,7 @@
 #'
 #' The dimensions and variables of the active grid are identified in the
 #' [print][print.tidync()] method of the tidync object, these functions exist to
-#' provide that information programmatically.
+#' provide that information directly.
 #'
 #' `hyper_vars()` will list the ids, data type, name, dimension number, number
 #' of attributes and and coordinate status of the variables on the currently
@@ -14,13 +14,16 @@
 #' 
 #' `hyper_dims()` will list the names, lengths, start/count index, ids, and
 #' status of dimensions on the currently active grid. records on the currently
-#' active dimensions
+#' active dimensions. 
+#' 
+#' `hyper_grids()` will list the names, number of dimension, and number of 
+#' variables and active status of each grid in the source.
 #' @param x tidync object
 #' @param ... ignored
 #'
 #' @return data frame
 #' @export
-#'
+#' @aliases hyper_dims hyper_grids
 #' @examples
 #' f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
 #' l3file <- system.file("extdata/oceandata", f, package= "tidync")
@@ -47,3 +50,9 @@ hyper_dims <- function(x, ...) {
   dplyr::inner_join(out, act1, c("length", "name"))
 }
 
+#' @name hyper_vars
+#' @export
+hyper_grids <- function(x, ...) {
+  x$grid[c("grid", "ndims", "nvars")] %>% 
+    mutate(active = .data$grid == active(x))
+}
