### 2019-03-18 #174 @mdsumner
I think I've addressed everything, thanks again for the responses!  I've done a run over the documentation and it's a lot tighter and uses links better now, every bit of feedback here has been really helpful and I appreciate the patience provided to me. I didn't foresee this being wrapped up so nicely a few weeks ago.

(For CRAN  it's down to needing a release of `ncmeta`, which I'm stuck on. )
