### 2019-03-18 #174 @mdsumner
It comes down to revdepcheck which I just can't get to finish ATM. I need to make sure a release doesn't disrupt stars,and I don't feel confident on that without revdep. If you revdepcheck ncmeta and share the results that would be helpful?
