### 2019-03-18 #174 @sckott
problems and readme from revdepcheck aattached

[problems.txt](https://github.com/ropensci/software-review/files/2980366/problems.txt)
[README.txt](https://github.com/ropensci/software-review/files/2980367/README.txt)
