### 2019-03-18 #174 @mdsumner
That was fast! Awesome, that's very helpful - I think I only need to condition on version pre and post and it'll  be safe. THANKS  (hopefully I can get revdep working again too ...)
