### 2019-03-20 #174 @mdsumner
I've got this sorted now, ncmeta will have to follow the next version of stars onto CRAN.
