### 2019-04-19 a299951
skip on cran some tests
