+++ man/tidync.Rd
@@ -1,13 +1,16 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tidync.R
+% Please edit documentation in R/tidync.R, R/tidync_data.R
 \name{tidync}
 \alias{tidync}
 \alias{tidync.character}
+\alias{tidync.tidync_data}
 \title{Tidy NetCDF}
 \usage{
 tidync(x, what, ...)
 
 \method{tidync}{character}(x, what, ...)
+
+\method{tidync}{tidync_data}(x, what, ...)
 }
 \arguments{
 \item{x}{path to a NetCDF file}
