+++ NAMESPACE
@@ -15,7 +15,9 @@ S3method(hyper_tibble,character)
 S3method(hyper_tibble,tidync)
 S3method(hyper_transforms,default)
 S3method(print,tidync)
+S3method(print,tidync_data)
 S3method(tidync,character)
+S3method(tidync,tidync_data)
 export("%>%")
 export("active<-")
 export(activate)
