### 2019-04-23 b2f9243
fixes for CRAN
