+++ R/magrittr-pipe.R
@@ -1,3 +1,6 @@
 #' @importFrom magrittr %>%
 #' @export
+#' @examples 
+#' system.file("extdata/argo/MD5903593_001.nc", package = "tidync") %>% 
+#'      tidync()
 magrittr::`%>%`
