+++ R/tidync-package.R
@@ -26,4 +26,16 @@
 #'  in printed output.
 #' @name tidync-package
 #' @docType package
+#' @examples 
+#' argofile <- system.file("extdata/argo/MD5903593_001.nc", package = "tidync")
+#' argo <- tidync(argofile)
+#' argo %>% active()
+#' argo %>% activate("D3,D8") %>% hyper_array()
+#' argo %>% hyper_filter(NLEVELS = index < 4)
+#' argo %>% hyper_tbl_cube()
+#' argo %>% hyper_tibble(select_var = c("TEMP_QC"))
+#' argo %>% hyper_transforms()
+#' argo %>% hyper_vars()
+#' argo %>% hyper_dims()
+#' argo %>% hyper_grids()
 NULL
