+++ man/deprecated-tidync.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/hyper_transforms.R
-\name{deprecated-tidync}
-\alias{deprecated-tidync}
-\alias{axis_transforms}
-\title{Deprecated functions in tidync}
-\usage{
-axis_transforms(x, ...)
-}
-\arguments{
-\item{x}{tidync object}
-
-\item{...}{ignored}
-}
-\description{
-Deprecated functions in tidync
-}
