+++ NEWS.md
@@ -1,8 +1,13 @@
-# tidync 0.1.0.9001
+# tidync 0.1.1
 
-* New class 'tidync_data' for output of 'hyper_array()', no underlying change to the
- object which is simply a list of arrays from each variable, and with the axis transforms
- stored in an attribute. 
+* Package improvements thanks to CRAN feedback, clarified Description and added
+ more examples. 
+
+* New class 'tidync_data' for output of 'hyper_array()', no underlying change 
+ to the object which is simply a list of arrays from each variable, and axis
+ transforms stored in an attribute. 
+
+* Old deprecated function 'axis_transforms()' now Defunct. 
 
 
 # tidync 0.1.0
