### 2019-04-26 5f782bd
remove print and cat from code that is not print-methods
