### 2019-04-27 #174 @mdsumner
And now on CRAN.

I still see stuff that needs fixing but functionally I think all is well. THANK YOU @sckott @timcdlucas and @Nowosad !   A while back I just didn't think this was going to work, but the whole process has been extremely helpful  and I appreciate all the patience.

It occurred to me I haven't credited efforts from this review enough in the package itself, but I will rectify that.
