### 2019-04-27 #174 @timcdlucas
Thanks, glad to have helped.

Didn't know print was explicitly banned. Will remember that.
