### 2019-04-29 f315e55
doc: read into R native arrays, not into raw arrays.
