+++ vignettes/netcdf-with-tidync.Rmd
@@ -23,7 +23,9 @@ options(pillar.subtle = FALSE, pillar.sigfig = 4)
 ```
 
 The goal of tidync is to ease exploring the contents of a NetCDF source and
-to easily extract array data in various forms.  Expressions relative to array dimensions are used to limit the data to be read, and this can be used interactively to explore the source data. No data is read until explicitly requested, and can be pulled directly as an array, or in long-form as a data frame. 
+to easily extract array data in various forms and automate complex extraction tasks. 
+
+Array dimension expressions virtually *slice* (i.e. *index*) into data arrays with immediate feedback, and ease programming for automating data extraction. No data is read until explicitly requested, and can be pulled directly as an array, or in long-form as a data frame. 
 
 NetCDF is **Network Common Data Form**, a [data model][not-a-format] and an
 [API](https://en.wikipedia.org/wiki/Application_programming_interface).
@@ -40,13 +42,12 @@ built-in files and online sources to demonstrate tidync functionality.
 NetCDF is a very widely used file format for storing array-based data as
 *variables*. The **space** occupied by a **variable** is defined by its
 **dimensions** and their metadata. Dimensions are by definition
-*one-dimensional* (i.e. an atomic vector, in R) consisting of one or more
-elements, a rectilinear virtual array with coordinate metadata on its units,
+*one-dimensional* (i.e. an atomic vector in R of length 1 or more), an array with coordinate metadata, units,
 type and interpretation. The **space** of a variable is defined as one or more
-of the dimensions in the file, but a variable won't necessarily use all the
+of the dimensions in the file. A given variable won't necessarily use all the
 available dimensions and no dimensions are mandatory or particularly special.
 
-Some conventions exist to define usage and minimal standards for metadata for particular file schemas, but these are many and varied, and not always adhered to. 
+Some conventions exist to define usage and minimal standards for metadata for particular file schemas, but these are many and varied, and not always adhered to. Tidync does not provide a data interpretation as it is intended for general use, including by tools that create data formats. 
 
 The R community is not particularly strong with use of NetCDF, though it
 is common and widely used it pales compared to use in general climate science work, and there the most used tool is the [CDO Climate Data Operators](https://code.mpimet.mpg.de/projects/cdo).  In R the most common
@@ -82,14 +83,11 @@ It's not particularly obvious that some variables are defined within the same
 set of dimensions as others.
 
 A NetCDF file is a container for simple array based data structures. There is
-[limited capacity][random-access] in the formal API for accessing data randomly
+limited capacity ^[I.e. it's not possible to query a file for an arbitrary sparse set of values, without constructing a degenerate hyperslab query for each point or reading a hyperslab containing cells not in the set.  Do you know different? Please let me know!] in the formal API for accessing data randomly
 within a variable, the primary mechanism is to define offset and stride (start
 and count) hyperslab indexes.
 
-[random-access]: I.e. it's not possible to query a file for an arbitrary sparse
-set of values, without constructing a degenerate hyperslab query for each point
-or reading a hyperslab containing cells not in the set.  Do you know different?
-Please let me know!
+
 
 ## tidync
 
@@ -183,7 +181,7 @@ The special 'dplyr' adverbs like 'between' will work.
 
 ```{r}
 concentration %>% 
-  hyper_filter(ni = index < 20, 
+  hyper_filter(ni = index < 50, 
                nj = dplyr::between(index, 30, 100))
 ```
 
[…]
