### 2019-05-10 #174 @mdsumner
@sckott are we good for rOpenSci here? I found some personal time to push this out to CRAN and kind of lost track of the status here -can we go to the next steps?
