### 2019-05-13 #174 @sckott
@mdsumner All is good, was waiting i think for ncmeta to get on CRAN, now it's on CRAN.

Approved!  Thanks again for your submission @mdsumner  ! And thanks for your reviews @Nowosad  and @timcdlucas  👌

To-dos:

* Please transfer the package  to `ropensci`- I've invited you to a team (you need to accept that invitation first, before transferring). You'll be made admin once you transfer
* add rOpenSci footer to README
`[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)`
* it's a good idea to add issue and PR template files to your .github folder, egs here https://github.com/ropensci/dotgithubfiles/
* We're starting to roll out software metadata files to all ropensci pkgs via the Codemeta initiative, see  https://github.com/ropensci/codemetar/#codemetar  for how to include it in your pkg, after installing the pkg - should be easy as running `codemetar::write_codemeta()` in the root of your pkg

---

We've started putting together a bookdown with our best practice and tips, [this chapter](https://ropensci.github.io/dev_guide/collaboration.html) starts the 3d section that's about guidance for after onboarding. Please tell us what could be improved. The repo is at https://github.com/ropensci/dev_guide

Are you interested in doing a blog post for our blog https://ropensci.org/blog/ ?  either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development (https://ropensci.org/blog/). If so, we'll have our community manager @stefaniebutland get in touch with you on that
