+++ man/hyper_array.Rd
@@ -38,7 +38,7 @@ applied or not}
 Extract the raw array data as a list of  one or more arrays. This can be the
 entire variable/s or after dimension-slicing using \code{\link[=hyper_filter]{hyper_filter()}}
 expressions. This is a delay-breaking function and causes data to be read
-from the source into raw arrays in R's native form. This list of arrays is
+from the source into R native arrays. This list of arrays is
 lightly classed as \link{tidync_data}, with methods for \code{\link[=print]{print()}} and \code{\link[=tidync]{tidync()}}.
 }
 \details{
