+++ tests/testthat/test-files.R
@@ -2,5 +2,5 @@ context("files")
 
 fname <- paste(sample(unlist(strsplit("somecrazyfile", ""))), collapse = "")
 test_that("file not found is friendly", {
-  expect_error(tidync(fname), "failed to open")
+  expect_error(expect_warning(tidync(fname)), "failed to open")
 })
