+++ README.Rmd
@@ -180,7 +180,12 @@ https://github.com/hypertidy/ncapi
 * **activation** - choice of a given grid to apply subsetting and read operations to
 
 
+---
+
 ## Code of conduct
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 
+
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
