### 2019-05-14 9d1d144
add rOpenSci templates
