+++ DESCRIPTION
@@ -33,10 +33,6 @@ Imports:
 RoxygenNote: 6.1.1
 Suggests: ggplot2,
     knitr,
-    memoise,
-    ncdump,
-    palr,
-    raster,
     rmarkdown,
     testthat,
     covr
