### 2019-05-14 4192100
remove unused purrr map
