+++ NAMESPACE
@@ -50,7 +50,6 @@ importFrom(ncdf4,nc_close)
 importFrom(ncdf4,nc_open)
 importFrom(ncdf4,ncvar_get)
 importFrom(ncmeta,nc_meta)
-importFrom(purrr,map)
 importFrom(purrr,safely)
 importFrom(rlang,.data)
 importFrom(stats,setNames)
