### 2019-05-14 81a1ce3
generalize links
