+++ README.Rmd
@@ -38,7 +38,7 @@ This project uses RNetCDF for the primary access to the NetCDF library, as well 
 
 Install tidync from CRAN. 
 
-```{r cran}
+```{r cran, eval = FALSE}
 install.packages("tidync")
 ```
 
@@ -89,7 +89,8 @@ install.packages("remotes")
 remotes::install_github("hypertidy/tidync")
 ```
 
-At the time of writing the [travis install configuration](https://github.com/hypertidy/tidync/blob/master/.travis.yml) was set up for "trusty", Ubuntu 14.04 so also provides a record. 
+At the time of writing the [travis install configuration](https://github.com/hypertidy/tidync/blob/master/.travis.yml) was set up for "xenial", Ubuntu 16.04 which was required for using [NetCDF sources with groups](https://www.unidata.ucar.edu/software/netcdf/workshops/2011/groups-types/GroupsIntro.html). 
+
 
 More general information about system dependencies libnetcdf-dev and libudunits2-dev is available from [Unidata NetCDF](https://www.unidata.ucar.edu/software/netcdf/docs/getting_and_building_netcdf.html) and [Unidata Udunits2](https://www.unidata.ucar.edu/software/udunits/udunits-current/doc/udunits/udunits2.html#Installation). 
 
