+++ man/hyper_tbl_cube.Rd
@@ -43,7 +43,7 @@ ufile <- system.file("extdata", "unidata", "test_hgroups.nc",
  package = "tidync", mustWork = TRUE)
  
 ## some versions of NetCDF don't support this file
-## (4.1.3 https://github.com/hypertidy/tidync/issues/82)
+## (4.1.3 tidync/issues/82)
 group_nc <- try(tidync(ufile), silent = TRUE)
 if (!inherits(group_nc, "try-error")) {
  res <-  hyper_tbl_cube(tidync(ufile))
