+++ README.Rmd
@@ -46,7 +46,7 @@ You can install the development version from github with:
 
 ```{r gh-installation, eval = FALSE}
 # install.packages("remotes")
-remotes::install_github("hypertidy/tidync", dependencies = TRUE)
+remotes::install_github("ropensci/tidync", dependencies = TRUE)
 ```
 
 The package packages ncdf4 and RNetCDF are required, so first make sure you can install and use these if it doesn't work the first time. 
@@ -86,10 +86,10 @@ Then in R
 
 ```R
 install.packages("remotes")
-remotes::install_github("hypertidy/tidync")
+remotes::install_github("ropensci/tidync")
 ```
 
-At the time of writing the [travis install configuration](https://github.com/hypertidy/tidync/blob/master/.travis.yml) was set up for "xenial", Ubuntu 16.04 which was required for using [NetCDF sources with groups](https://www.unidata.ucar.edu/software/netcdf/workshops/2011/groups-types/GroupsIntro.html). 
+At the time of writing the [travis install configuration](https://github.com/ropensci/tidync/blob/master/.travis.yml) was set up for "xenial", Ubuntu 16.04 which was required for using [NetCDF sources with groups](https://www.unidata.ucar.edu/software/netcdf/workshops/2011/groups-types/GroupsIntro.html). 
 
 
 More general information about system dependencies libnetcdf-dev and libudunits2-dev is available from [Unidata NetCDF](https://www.unidata.ucar.edu/software/netcdf/docs/getting_and_building_netcdf.html) and [Unidata Udunits2](https://www.unidata.ucar.edu/software/udunits/udunits-current/doc/udunits/udunits2.html#Installation). 
