### 2019-05-14 #174 @mdsumner
Great thanks!  All done apart from

- [ ] admin for me on ropensci/tidync (change pkgdown doc link)
- [ ] appveyor badge

tidync 0.2.0 just hit CRAN!
