### 2019-05-14 #174 @stefaniebutland
Most excellent!
Guidelines are here: https://github.com/ropensci/roweb2#contributing-a-blog-post and worth checking as we've made some updates. Earliest publication date would be June 11, 2019. If that works for you, please submit a draft via pull request by June 4, 2019.

Once published, it will show up on [your author page](https://ropensci.org/authors/michael-sumner/)

Happy to answer any questions.
