+++ .travis.yml
@@ -44,3 +44,4 @@ notifications:
 
 after_success:
   - Rscript -e 'covr::codecov()'
+  
