### 2019-05-15 ac715d4
appveyor redirect link, and codemeta flush
