### 2019-05-15 #174 @sckott
sorry to change recommendation @mdsumner - we recommend continuing to maintain your appveyor badge under your own acct
