### 2019-05-15 #174 @mdsumner
Oh that's all good, the change is only that appveyor creates a new link (another random suffix), but the badge hangs off github/ropensci:

```
[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ropensci/tidync?branch=master&svg=true)](https://ci.appveyor.com/project/mdsumner/tidync-nvwwt)
```
