### 2019-05-15 b454275
fix hyper_array axis order problem, and adjusted tests to suit
